+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * MI-164: set nginx logrotate to only keep 30 
 * MI-171: nginx config reload needs to watch for changes to both the ssl key *and* cert
+* MI-161: Install latest CrowdStrike Falcon Sensor package
 
 ## v2.9 - 7/19/2019